requirements.txt
django==4.2
psycopg2-binary==2.9.10
pillow==10.4.0
python-dotenv==1.0.1
djangorestframework==3.15.2
drf-yasg==1.21.8
